# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Kaay entreprendre",
        "description": "Cet api va nous permettre de faire communiquer la partie front et la partie back de notre plateforme",
        "version": "1.0.0"
    },
    "components": {
        "securitySchemes": {
          "BearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        }
      },
      "security": [
        {
          "BearerAuth": []
        }
      ], 
    "paths": {
...
        "/api/commentaire": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Lister un commentaire d'un forum",
                "operationId": "ce316673403fb864ec9b6a97e81af53e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Détails du commentaire"
                    }
                }
            }
        },
        "/api/commentaire/create": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Ajouter un commentaire à un forum",
                "operationId": "4350d9aff5913143561dfac9f41cd741",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "contenu": {
                                        "type": "string",
                                        "example": "Contenu du commentaire"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Le commentaire est bien ajouté"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/commentaire/edit": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Modifier un commentaire d'un forum",
                "operationId": "1c089d0190e1d533568040e5ce06b74e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "contenu": {
                                        "type": "string",
                                        "example": "Nouveau contenu du commentaire"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Le commentaire est bien modifié"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/commentaire/archive": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Archiver un commentaire de forum",
                "operationId": "9df982e2ee36e35716538192fedf7a5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Le commentaire est bien archivé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/commentaire/delete": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Supprimer un commentaire d'un forum",
                "operationId": "03d04fad5791d252a4200ac9f7238190",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Le commentaire est bien supprimé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
...
    },
...
}